DEV: Fix performance of group posts query (#3)
Noticed a performance issue with the group posts query after deploying
the plugin to a medium-sized site. The issue is that the mix of posts *
group_users * tags was making Postgres load every row in posts.

Thanks to @xfalcox for the solution implemented here: splitting the
complex query into two simpler queries and merging the results.